split event manager and event publisher
...in order to resolve the circular dependencies, caused by
many EventSubscriber implementation also requiring an EventPublisher.

The service tracker which was included in the OSGiEventManager was
just attempting to hide these circular dependencies from the SCR,
but still caused some trouble sometimes.

fixes #4356
fixes #4775
